# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/shelf-click:
    post:
      operationId: recordShelfClicks
      tags: [analytics]
      summary: Record shelf outbound clicks
      description: |
        Persist a batch of "Visit brand" outbound clicks emitted by the
        "On our shelf" cards on ingredient briefs, routine pages, and
        concern guides. Designed to be called from
        `navigator.sendBeacon` (or `fetch` with `keepalive`); the
        endpoint always responds 204 so the browser never blocks
        navigation waiting on a body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShelfClickBatch"
      responses:
        "204":
          description: Accepted — clicks were persisted.
        "400":
          description: Malformed payload.
  /analytics/shelf-clicks:
    get:
      operationId: listShelfClicks
      tags: [analytics]
      summary: List recent shelf outbound clicks
      description: |
        Editorial-only readout of recent "Visit brand" clicks, newest
        first. Returns at most `limit` rows (default 200, max 1000) so
        editors can spot which brand links readers actually use.

        Optional `since`/`until` ISO-8601 timestamps narrow the window
        in SQL — necessary so date-range chips like `Last 30 days` or
        `All recent` don't silently cap at the most recent 1000 rows
        once a busy month exceeds that ceiling.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
        - name: pageKind
          in: query
          required: false
          description: Optional filter to a single page kind (ingredient, routine, concern).
          schema:
            $ref: "#/components/schemas/ShelfPageKind"
        - name: pageSlug
          in: query
          required: false
          description: |
            Optional filter to a single source page (e.g. an ingredient
            slug like `tretinoin`, or a routine/concern component name
            like `RoutineBareMinimum`).
          schema:
            type: string
        - name: since
          in: query
          required: false
          description: |
            Inclusive lower bound on `createdAt` (ISO-8601). When set,
            only clicks with `createdAt >= since` are returned.
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          required: false
          description: |
            Inclusive upper bound on `createdAt` (ISO-8601). When set,
            only clicks with `createdAt <= until` are returned.
          schema:
            type: string
            format: date-time
      security:
        - editorSession: []
      responses:
        "200":
          description: Recent shelf clicks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShelfClickList"
        "401":
          description: |
            Missing or invalid editor session. Callers must first sign
            in via `POST /editor/sign-in` to receive the HttpOnly
            `evidently_editor` session cookie.
  /analytics/shelf-clicks/export:
    get:
      operationId: exportShelfClicks
      tags: [analytics]
      summary: Stream every shelf click in a date window as NDJSON
      description: |
        Editorial-only export endpoint for "Download CSV". Unlike the
        dashboard's `GET /analytics/shelf-clicks` listing — which is
        capped at 1000 rows so a busy month silently truncates — this
        endpoint streams every row matching the supplied filters, one
        JSON object per line (`application/x-ndjson`). The dashboard
        builds the CSV from the stream so the column order, escaping,
        and page-label resolution stay identical to the in-memory
        exporter.

        The stream is keyset-paginated server-side on
        `(createdAt DESC, id DESC)` so only one page is in flight at a
        time and memory / database connection use stays bounded
        regardless of how busy the window is. There is no upper row
        cap — the whole point of the route is to give editors the full
        date window without silent truncation, so the only bounds are
        the `since` / `until` query params.
      parameters:
        - name: pageKind
          in: query
          required: false
          description: Optional filter to a single page kind (ingredient, routine, concern).
          schema:
            $ref: "#/components/schemas/ShelfPageKind"
        - name: pageSlug
          in: query
          required: false
          description: |
            Optional filter to a single source page (e.g. an ingredient
            slug like `tretinoin`, or a routine/concern component name
            like `RoutineBareMinimum`). Used by the dashboard's drill-in
            view so the export only contains the page the editor is
            looking at.
          schema:
            type: string
        - name: since
          in: query
          required: false
          description: |
            Inclusive lower bound on `createdAt` (ISO-8601). When set,
            only clicks with `createdAt >= since` are streamed.
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          required: false
          description: |
            Inclusive upper bound on `createdAt` (ISO-8601). When set,
            only clicks with `createdAt <= until` are streamed.
          schema:
            type: string
            format: date-time
      security:
        - editorSession: []
      responses:
        "200":
          description: |
            Newline-delimited stream of `ShelfClickRecord` objects
            (one JSON object per line, no enclosing array). Newest
            first, matching the dashboard's listing order. The
            response may be empty (no rows in the window) but a
            successful export still returns 200.
          content:
            application/x-ndjson:
              schema:
                type: string
                description: |
                  One `ShelfClickRecord` JSON object per line, terminated
                  by `\n`. Clients should parse line-by-line rather than
                  buffering the whole body, since busy windows can be
                  arbitrarily large.
        "400":
          description: Malformed query string.
        "401":
          description: |
            Missing or invalid editor session. Callers must first sign
            in via `POST /editor/sign-in`.
  /editor/sign-in:
    post:
      operationId: editorSignIn
      tags: [analytics]
      summary: Exchange the shared editor token for a session cookie
      description: |
        The editor dashboard prompts for the shared `EDITOR_TOKEN`
        once. Posting it here exchanges it for an HttpOnly,
        signed session cookie (`evidently_editor`) that the
        editor-only API surface (currently `GET /analytics/shelf-clicks`)
        validates. The token itself is never returned to the browser
        and the cookie cannot be read from JavaScript.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorSignInBody"
      responses:
        "204":
          description: Session cookie issued.
          headers:
            Set-Cookie:
              description: HttpOnly signed `evidently_editor` cookie.
              schema:
                type: string
        "401":
          description: Token did not match the configured editor token.
        "503":
          description: |
            Server-side `EDITOR_TOKEN` is not configured, so no editor
            credential can possibly be valid.
...
components:
  securitySchemes:
    editorSession:
      type: apiKey
      in: cookie
      name: evidently_editor
      description: |
        HttpOnly signed session cookie. Issued by `POST /editor/sign-in`
        in exchange for the shared `EDITOR_TOKEN`, and valid for 30
        days. Endpoints without this scheme listed are public.
  schemas:
...
    EditorSignInBody:
      type: object
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 1000
          description: |
            The shared editor token (the value of the server's
            `EDITOR_TOKEN` env var). Sent only on this request and never
            echoed back to the browser.
      required:
        - token
...
    ShelfPageKind:
      type: string
      enum: [ingredient, routine, concern]
      description: |
        Which kind of editorial page surfaced the shelf card. Combined
        with `pageSlug`, this identifies the source page so analytics
        can be sliced by page kind, by individual page, or both.
    ShelfClickEvent:
      type: object
      properties:
        pageKind:
          $ref: "#/components/schemas/ShelfPageKind"
        pageSlug:
          type: string
          minLength: 1
          maxLength: 200
          description: |
            Per-kind identifier for the source page. Ingredient briefs
            use the catalogue ingredient slug (e.g. `tretinoin`); routine
            and concern pages use the detail-page component file name
            (e.g. `RoutineBareMinimum`, `ConcernRosacea`) because
            multiple catalogue slugs can alias the same component.
        brand:
          type: string
          minLength: 1
          maxLength: 200
        productName:
          type: string
          minLength: 1
          maxLength: 400
        href:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - pageKind
        - pageSlug
        - brand
        - productName
        - href
    ShelfClickBatch:
      type: object
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/ShelfClickEvent"
      required:
        - events
    ShelfClickRecord:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        pageKind:
          $ref: "#/components/schemas/ShelfPageKind"
        pageSlug:
          type: string
        brand:
          type: string
        productName:
          type: string
        href:
          type: string
        userAgent:
          type: string
          nullable: true
      required:
        - id
        - createdAt
        - pageKind
        - pageSlug
        - brand
        - productName
        - href
    ShelfClickList:
      type: object
      properties:
        clicks:
          type: array
          items:
            $ref: "#/components/schemas/ShelfClickRecord"
        limit:
          type: integer
          description: |
            The row cap that was applied to this query — i.e. the
            maximum number of rows the response could possibly contain.
            Returned so clients can phrase a "showing the most recent N"
            banner using the same number the server actually used.
        hasMore:
          type: boolean
          description: |
            `true` when matching rows exist past the response — i.e.
            when `totalCount > limit`. The dashboard uses this to warn
            editors that the totals on screen are a truncated view and
            they should narrow the date range to see the full picture.
            Computed from the count rather than from `clicks.length ===
            limit` so the exact-edge case where the total equals the
            cap (`totalCount === limit`) does not falsely claim
            truncation.
        totalCount:
          type: integer
          minimum: 0
          description: |
            Total number of rows matching the same `pageKind` / `pageSlug`
            / `since` / `until` filters, ignoring `limit`. Always present
            so clients can phrase the truncation concretely (e.g.
            "Showing the most recent 1,000 of 12,438 clicks"). When the
            cap was not hit the server avoids the extra `SELECT COUNT(*)`
            round-trip and reports `clicks.length` directly — that
            already is the true total in that case, so this field is
            cheap on the common path and only triggers a count query
            when `hasMore === true`.
      required:
        - clicks
        - limit
        - hasMore
        - totalCount
...